refactor: flatten query folder, services folder
