+++ apps/app/src/lib/query/manualRecorder.ts
@@ -1,24 +1,24 @@
 import * as services from '$lib/services';
 import { toast } from '$lib/toast';
 import type { WhisperingRecordingState } from '$lib/constants';
-import { defineMutation, defineQuery } from '../_utils';
-import { queryClient } from '../index';
+import { defineMutation, defineQuery } from './_utils';
+import { queryClient } from './index';
 import { settings } from '$lib/stores/settings.svelte';
 
 const recorderKeys = {
-	state: ['cpalRecorder', 'state'] as const,
-	startRecording: ['cpalRecorder', 'startRecording'] as const,
-	stopRecording: ['cpalRecorder', 'stopRecording'] as const,
-	cancelRecording: ['cpalRecorder', 'cancelRecording'] as const,
+	state: ['recorder', 'state'] as const,
+	startRecording: ['recorder', 'startRecording'] as const,
+	stopRecording: ['recorder', 'stopRecording'] as const,
+	cancelRecording: ['recorder', 'cancelRecording'] as const,
 } as const;
 
 const invalidateRecorderState = () =>
 	queryClient.invalidateQueries({ queryKey: recorderKeys.state });
 
-export const cpalRecorder = {
+export const manualRecorder = {
 	getRecorderState: defineQuery({
 		queryKey: recorderKeys.state,
-		resultQueryFn: () => services.cpalRecorder.getRecorderState(),
+		resultQueryFn: () => services.manualRecorder.getRecorderState(),
 		initialData: 'IDLE' as WhisperingRecordingState,
 	}),
 
@@ -26,20 +26,20 @@ export const cpalRecorder = {
 		mutationKey: recorderKeys.startRecording,
 		resultMutationFn: ({
 			toastId,
-			selectedDeviceId,
+			recordingSettings,
 		}: {
 			toastId: string;
-			selectedDeviceId: string | null;
+			recordingSettings: {
+				selectedDeviceId: string | null;
+				bitrateKbps: string;
+			};
 		}) => {
-			if (settings.value['recording.mode'] !== 'cpal') {
-				settings.value = { ...settings.value, 'recording.mode': 'cpal' };
+			if (settings.value['recording.mode'] !== 'manual') {
+				settings.value = { ...settings.value, 'recording.mode': 'manual' };
 			}
-			return services.cpalRecorder.startRecording(
-				{ selectedDeviceId },
-				{
-					sendStatus: (options) => toast.loading({ id: toastId, ...options }),
-				},
-			);
+			return services.manualRecorder.startRecording(recordingSettings, {
+				sendStatus: (options) => toast.loading({ id: toastId, ...options }),
+			});
 		},
 		onSettled: invalidateRecorderState,
 	}),
@@ -47,7 +47,7 @@ export const cpalRecorder = {
 	stopRecording: defineMutation({
 		mutationKey: recorderKeys.stopRecording,
 		resultMutationFn: ({ toastId }: { toastId: string }) =>
-			services.cpalRecorder.stopRecording({
+			services.manualRecorder.stopRecording({
 				sendStatus: (options) => toast.loading({ id: toastId, ...options }),
 			}),
 		onSettled: invalidateRecorderState,
@@ -56,7 +56,7 @@ export const cpalRecorder = {
 	cancelRecording: defineMutation({
 		mutationKey: recorderKeys.cancelRecording,
 		resultMutationFn: ({ toastId }: { toastId: string }) =>
-			services.cpalRecorder.cancelRecording({
+			services.manualRecorder.cancelRecording({
 				sendStatus: (options) => toast.loading({ id: toastId, ...options }),
 			}),
 		onSettled: invalidateRecorderState,
