+++ apps/app/src/lib/query/recordings.ts
@@ -2,8 +2,8 @@ import type { Recording } from '$lib/services/db';
 import * as services from '$lib/services';
 import { Err, Ok } from '@epicenterhq/result';
 import type { Accessor } from '@tanstack/svelte-query';
-import { defineMutation, defineQuery } from '../_utils';
-import { queryClient } from '../index';
+import { defineMutation, defineQuery } from './_utils';
+import { queryClient } from './index';
 
 const recordingKeys = {
 	all: ['recordings'] as const,
