+++ apps/app/src/lib/query/transformer.ts
@@ -5,8 +5,8 @@ import type {
 } from '$lib/services/db';
 import { Err, Ok, type Result } from '@epicenterhq/result';
 import type { WhisperingError, WhisperingResult } from '$lib/result';
-import { defineMutation } from '../_utils';
-import { queryClient } from '../index';
+import { defineMutation } from './_utils';
+import { queryClient } from './index';
 import { transformationRunKeys } from './transformationRuns';
 import { transformationsKeys } from './transformations';
