+++ apps/app/src/lib/query/vadRecorder.ts
@@ -2,8 +2,8 @@ import * as services from '$lib/services';
 import { settings } from '$lib/stores/settings.svelte';
 import { Ok } from '@epicenterhq/result';
 import type { VadState } from '$lib/constants';
-import { defineMutation, defineQuery } from '../_utils';
-import { queryClient } from '../index';
+import { defineMutation, defineQuery } from './_utils';
+import { queryClient } from './index';
 
 const vadRecorderKeys = {
 	all: ['vadRecorder'] as const,
