+++ apps/app/src/lib/services/transcription/speaches.ts
@@ -3,7 +3,7 @@ import { getExtensionFromAudioBlob } from '$lib/utils';
 import { Err, Ok, type Result } from '@epicenterhq/result';
 import type { WhisperingError } from '$lib/result';
 import { z } from 'zod';
-import type { TranscriptionService, TranscriptionServiceError } from '..';
+import type { TranscriptionService, TranscriptionServiceError } from '.';
 
 const whisperApiResponseSchema = z.union([
 	z.object({ text: z.string() }),
